+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Removed resource constraints on backend and frontend.
+- Epinio is no longer installed.
 
 ## [1.0.4] - 2024-12-19